+++ packages/executor/src/types/common.ts
@@ -1,3 +1,0 @@
-export type Time = {
-	getTimestamp(): Promise<number>;
-};