+++ packages/gateways/src/KVSchedulerStorage.ts
@@ -8,16 +8,9 @@ function lexicographicNumber(num: number, size: number): string {
 	return num.toString().padStart(size, '0');
 }
 
-function computeQueueID(
-	checkinTime: number,
-	chainId: `0x${string}`,
-	id: string,
-	timeContract: `0x${string}` | undefined
-): string {
-	const prefix = timeContract ? `q_${chainId}_${timeContract}` : 'q__';
-	const suffix = timeContract
-		? `${lexicographicNumber(checkinTime, 12)}_${id}`
-		: `${lexicographicNumber(checkinTime, 12)}_${chainId}_${id}`;
+function computeQueueID(checkinTime: number, chainId: `0x${string}`, id: string): string {
+	const prefix = 'q__';
+	const suffix = `${lexicographicNumber(checkinTime, 12)}_${chainId}_${id}`;
 	return `${prefix}${suffix}`;
 }
 
@@ -28,67 +21,43 @@ export class KVSchedulerStorage<TransactionDataType> implements SchedulerStorage
 		chainId: `0x${string}`;
 		id: string;
 		checkinTime: number;
-		timeContract?: `0x${string}`;
 	}): Promise<ExecutionQueued<TransactionDataType> | undefined> {
 		return this.db.get<ExecutionQueued<TransactionDataType>>(
-			computeQueueID(params.checkinTime, params.chainId, params.id, params.timeContract)
+			computeQueueID(params.checkinTime, params.chainId, params.id),
 		);
 	}
-	async deleteExecution(params: {
-		chainId: `0x${string}`;
-		timeContract?: `0x${string}`;
-		id: string;
-		checkinTime: number;
-	}): Promise<void> {
-		await this.db.delete(computeQueueID(params.checkinTime, params.chainId, params.id, params.timeContract));
+	async deleteExecution(params: {chainId: `0x${string}`; id: string; checkinTime: number}): Promise<void> {
+		await this.db.delete(computeQueueID(params.checkinTime, params.chainId, params.id));
 	}
 	async queueExecution(
-		executionToStore: ExecutionQueued<TransactionDataType>
+		executionToStore: ExecutionQueued<TransactionDataType>,
 	): Promise<ExecutionQueued<TransactionDataType>> {
 		await this.db.put<ExecutionQueued<TransactionDataType>>(
-			computeQueueID(
-				executionToStore.checkinTime,
-				executionToStore.chainId,
-				executionToStore.id,
-				executionToStore.timeContract
-			),
-			executionToStore
+			computeQueueID(executionToStore.checkinTime, executionToStore.chainId, executionToStore.id),
+			executionToStore,
 		);
 		return executionToStore;
 	}
 	async updateExecutionInQueue(executionUpdated: ExecutionQueued<TransactionDataType>): Promise<void> {
 		await this.db.put(
-			computeQueueID(
-				executionUpdated.checkinTime,
-				executionUpdated.chainId,
-				executionUpdated.id,
-				executionUpdated.timeContract
-			),
-			executionUpdated
+			computeQueueID(executionUpdated.checkinTime, executionUpdated.chainId, executionUpdated.id),
+			executionUpdated,
 		);
 	}
 	async reassignExecutionInQueue(
 		oldExecutionTime: number,
-		execution: ExecutionQueued<TransactionDataType>
+		execution: ExecutionQueued<TransactionDataType>,
 	): Promise<void> {
 		await this.db.transaction(async (txn) => {
-			await txn.delete(computeQueueID(oldExecutionTime, execution.chainId, execution.id, execution.timeContract));
+			await txn.delete(computeQueueID(oldExecutionTime, execution.chainId, execution.id));
 			await txn.put<ExecutionQueued<TransactionDataType>>(
-				computeQueueID(execution.checkinTime, execution.chainId, execution.id, execution.timeContract),
-				execution
+				computeQueueID(execution.checkinTime, execution.chainId, execution.id),
+				execution,
 			);
 		});
 	}
-	async getQueueTopMostExecutions(
-		params: {limit: number},
-		onlyWithTimeContract?: {
-			chainId: `0x${string}`;
-			timeContract: `0x${string}`;
-		}
-	): Promise<ExecutionQueued<TransactionDataType>[]> {
-		const prefix = onlyWithTimeContract
-			? `q_${onlyWithTimeContract.chainId}_${onlyWithTimeContract.timeContract}`
-			: 'q__';
+	async getQueueTopMostExecutions(params: {limit: number}): Promise<ExecutionQueued<TransactionDataType>[]> {
+		const prefix = 'q__';
 		const map = await this.db.list<ExecutionQueued<TransactionDataType>>({prefix, limit: params.limit});
 		const values: ExecutionQueued<TransactionDataType>[] = [];
 		for (const value of map.values()) {