+++ platforms/cf-worker/src/SchedulerDO.ts
@@ -9,14 +9,7 @@ import {
 	BroadcasterSignerData,
 } from 'fuzd-executor';
 import {KVExecutorStorage, KVSchedulerStorage, initSchedulerGateway} from 'fuzd-gateways';
-import {
-	ChainConfigs,
-	Scheduler,
-	SchedulerBackend,
-	SchedulerStorage,
-	WithTimeContract,
-	createScheduler,
-} from 'fuzd-scheduler';
+import {ChainConfigs, Scheduler, SchedulerBackend, SchedulerStorage, createScheduler} from 'fuzd-scheduler';
 import {initDecrypter} from 'fuzd-tlock-decrypter';
 
 import {JSONRPCHTTPProvider} from 'eip-1193-json-provider';
@@ -167,8 +160,8 @@ export class SchedulerDO extends createDurable() {
 		return new Response(this.account.publicExtendedKey);
 	}
 
-	processQueue(onlyWithTimeContract?: WithTimeContract) {
-		return this.scheduler.processQueue(onlyWithTimeContract);
+	processQueue() {
+		return this.scheduler.processQueue();
 	}
 
 	processPendingTransactions() {