+++ packages/executor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fuzd-executor
 
+## 0.1.4
+
+### Patch Changes
+
+- remove timeContract tracking
+- Updated dependencies
+  - fuzd-common@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes